Rename v1.py to 3 Different Mapping Approaches.py
Programmer: Ali Bukhari

Date: August 6th, 2025

Purpose: To try three different methods of allocating DA-level quintile scores on each of the four ON-MARG dimensions to an FSA.

Inputs: 
1. DA-level ON-MARG scores and quintiles from ON-MARG+
2. DA-to-FSA mapping file produced in DA_to_FSA_Overlap_Mapping in same repository
3. DA pop counts from 2021 Census

Method 1 (area-based): Each DA is assigned the FSA of which the majority of its area is in. Then, the number of DA in each quintile in each dimension in each FSA are counted and converted to a % of that FSA's area

Method 2 (population-based): Each DA and its population are assigned the FSA of which the majority of its area is in. Then, the population of DA in each quintile in each dimension are summed at the FSA level and converted to % of that FSA

Method 3 (weighted population-based): Each DA's population is allocated to an FSA on the basis of the count of postal codes of that DA present in that FSA. E.g., if DA X has 10 postal codes in its area, and 8 of them are in L0C and 2 in L9P, its population will be split 80%-20% between those FSAs. The rest of the steps are the sum as method 2.